requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
httpx==0.25.1
pydantic==2.4.2
cachetools==5.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
aiohttp==3.9.1
aiodns==3.1.1
ujson==5.8.0
